Move query.js to metabase/lib
